move the content_negotiation middleware into rack folder and  module
